Split layout to server and client sides
